feat: read / non-read 로 service 구분
